Use uint instead of int as property type for some entities. Update ORM lib to 0.0.14.76 (Fixes enum to T conversion crashes)

+++ Projects/Framework/Database/Data/Entities/NameGen.cs
@@ -22,9 +22,9 @@ namespace Framework.Database.Data.Entities
 {
     public class NameGen : Entity
     {
-        public int Id      { get; set; }
-        public string Name { get; set; }
-        public Race RaceId { get; set; }
-        public int Sex     { get; set; }
+        public uint Id      { get; set; }
+        public string Name  { get; set; }
+        public Race RaceId  { get; set; }
+        public byte Sex     { get; set; }
     }
 }
